[BEAM-12591] move SchemaHelpers to correct package
